Move `LoggingService::list_logs` to a separate module
